[Design] : 하단 탭바 배치 #1
