+++ GitHubSearchApp_iOS_TCA/Sources/GitHubSearchAppComponent/MyRepo/Views/MyRepoView.swift
@@ -1,0 +1,14 @@
+//
+//  MyRepoView.swift
+//  GitHubSearchApp_iOS_TCA
+//
+//  Created by minii on 2023/03/06.
+//
+
+import SwiftUI
+
+struct MyRepoView: View {
+  var body: some View {
+    Text("MyRepoView")
+  }
+}
